+++ NEWS.md
@@ -7,6 +7,9 @@
 ## `SCpubr::do_CellularStatesPlot()`
 -   Fixed a bug that prevented FeaturePlots to have symmetrical axes with respect to the main plot.
 
+## `SCpubr::do_EnrichmentHeatmap()`
+-   Fixed a bug in the code that prevented the feature plots and the geyser plots to be computed if the input gene list was not a named list of genes.
+
 # SCpubr v1.0.4
 -   Hotfix for v1.0.3 in which `do_GeyserPlot` with categorical variables had a bug that mapped the legend to the continuous axis.